Add files via upload
The "WebScrapeSearchGUI" project is a Python application designed to provide users with a graphical interface for searching keywords within the content of a web page obtained through web scraping. Utilizing the BeautifulSoup library for parsing HTML and Tkinter for creating the graphical user interface, the application allows users to input a URL and a list of keywords, which are then used to scrape the web page's content. The application handles exceptions gracefully, providing informative error messages in case of network issues or unexpected errors. Search results are displayed within the GUI, with keywords highlighted for easy identification, enhancing the user's experience in exploring and analyzing web content efficiently.